commit Amélioration des test IPokemonFactoryTest.java

+++ src/main/java/fr/univavignon/pokedex/api/PokemonFactory.java
@@ -21,7 +21,7 @@ public class PokemonFactory implements IPokemonFactory{
             int attack = rand.nextInt(16);
             int def = rand.nextInt(16);
             int stami = rand.nextInt(16);
-            double iv =  ((attack + def + stami) * 100 / 45);
+            int iv = (int)((attack + def + stami) * 100 / 45);
 
             Pokemon pokemon = new Pokemon(index, pokemonMetadata.getName(), pokemonMetadata.getAttack()+attack, pokemonMetadata.getDefense()+def, pokemonMetadata.getStamina()+stami, cp, hp, dust, candy, iv);
             return pokemon;
